+++ app/helpers/pro_connect_helper.rb
@@ -55,7 +55,7 @@ module ProConnectHelper
       user_info = get_user_info(access_token)
       return false if user_info.blank?
 
-      cree_ou_recupere_compte(user_info)
+      ProConnectRecupereCompteHelper.cree_ou_recupere_compte(user_info)
     end
 
     def recupere_tokens(code, callback_url)
@@ -80,41 +80,6 @@ module ProConnectHelper
       return false unless res.success?
 
       decode_jwt(res.body) { |user_info| user_info }
-    end
-
-    def cree_ou_recupere_compte(user_info)
-      email = user_info['email'].strip.downcase
-      compte = Compte.find_by(id_pro_connect: user_info['sub'])
-      if compte.present? && compte.email != email
-        compte = actualise_email_compte_existant(compte, email)
-      end
-      compte ||= Compte.find_or_create_by(email: email)
-      actualise_autres_champs(compte, user_info)
-      compte.save!
-      compte
-    end
-
-    private
-
-    def actualise_email_compte_existant(compte, email)
-      compte_existant = Compte.find_by(email: email)
-      if compte_existant.present?
-        compte.update!(id_pro_connect: nil)
-        compte = compte_existant
-      else
-        compte.email = email
-        compte.skip_reconfirmation!
-        compte.confirmed_at = Time.zone.now
-      end
-      compte
-    end
-
-    def actualise_autres_champs(compte, user_info)
-      compte.id_pro_connect = user_info['sub']
-      compte.prenom = user_info['given_name']
-      compte.nom = user_info['usual_name']
-      compte.password = SecureRandom.uuid if compte.encrypted_password.blank?
-      compte.confirmed_at ||= Time.zone.now
     end
   end
 end
